Email authentication is now required for users on sign up. If they attempt to login to an account with an email which has not been authenticated, they are taken to a screen where they can then resend the authentication email.
The bug where duplicate wallets would appear has also been fixed, the issue was due to the wallets being fetched by two seperate functions.

Added an option on the main menu screen where users can add their phone number in order to sign in with SMS authentication. I want the user to be required to enter an SMS code whenever they sign in if they have SMS authentication enabled, however this currently is not working, as users can sign in normally even if they have a phone number linked to their account. Will focus on implementing this correctly before the next commit.

In addition, I have edited the code for the sign up page, as previously I was hashing the user passwords on the app and then saving them to the database, however this is unnecessary as google firebase automatically hashes passwords.

A minor change is that most of the comments from the kotlin files have been removed, as I felt that they were unnecessary and caused too much clutter.