Edited the code handling SMS authentication. The user is now correctly enrolled into MFA when they enter the phone number, and enter the verification code. Previously, due to my lack of knowledge surrounding the implementation of MFA in firebase, the code just linked the phone number to the user's account, rather than correctly enrolling them in SMS MFA.
Now, whenever a user signs in, the app communicates with firebase to check if the user is enrolled in SMS MFA. If they are, they are taken to a page where they enter the verification code which was sent to their device, and if the code is valid, they are signed in to the app. This adds another layer of security to the app.