Some of the final updates have been made in this commit.
Two new features were added:
Password reset- On the sign up page, there is a button which takes the user to a screen where they can enter their email address and have a password reset link sent to them via Firebase Authentication, finally allowing users to update passwords.

Account deletion- On the main menu, there's a new option for account deletion, which will take users to a page where they can choose to delete the account from the database. This action will remove all of the seed phrase entries which contained the userID, as well as removing the user from the users collection on the database. Their account will also be removed permanently from Firebase Authentication.

In addition to these changes, a minor bug was fixed, previously, when entering a new seed phrase, the name of the wallet could be left blank. I've now updated this, so theres a new function on the AddNewSeedPhraseActivity class which checks if the text box is empty, and prompts the user to enter the name of the wallet.

The UI has also been overhauled, making it more in line with the original Figma designs. Most of the buttons now have the same colour and text font. The input boxes across the app now use the same style, the "@android:drawable/editbox_background" appearance. In addition, on most of the layout files, the gravity of the content is centered, meaning the app will have a more consistent UI across different screen sizes. The themes.xml file has also been updated, meaning that there is no longer an action bar at the top of the screen displaying the name of the activity, giving the app a more minimalist and aesthetically pleasing UI.