A patch fixing a bug with duplicate wallet entries.
Previously, if a user created two wallets with the same name, upon viewing each of the wallets, the list of phrases would be identical for both. This was because the entries from the database were fetched using the name of the wallet, rather than the documentID.

This patch addresses this issue, by using the documentID for crud operations, and making queries with document() rather than whereEqualTo, providing direct access and also making the code more efficient.