Fixed a bug which allowed for duplicate wallet entries to be made if the submit button was spammed.
To fix this, I added button debouncing to prevent multiple submissions being made if spammed by the user. This was done by implementing a lock mechanism using a boolean flag "isSubmitting". The flag starts as false, allowing for a submission to be made, and is then set to true when submission begins, preventing additional clicks. The flag is reset to false after a failed or successful save to the database.