Fixed the logic for enrolling the user in SMS 2FA.
Previous testing of the SMS 2FA feature involved using test numbers from the firebase console, which have a preset number and verification code. However, upon attempting to test the feature using my own phone number, I got an error saying that the code had expired, despite entering the code almost instantly.

After reviewing the official Firebase documentation, I found out that the issue was due to the MultiFactorSession object not being included in the function for sending the verification codes, which would cause the codes to instantly expire.

To fix this, I updated the sendVerificationCode function to properly pass the MultiFactorSession object to PhoneAuthOptions, fixing the issue and correctly sending the code. I also added null checks for the session before verification; improved the error handling on the verifyCode function; and removed the function to save the phone number to Firestore, as it is unnecessary to store the phone numbers there in order for SMS 2FA to function.