The option on the main menu for enabling SMS 2FA is now hidden if the user is already enrolled with a phone number.
To do this, I simply added a new function which would make a request to Firebase Auth, retrieving the user's enrolled factors and checking to see whether they were enrolled with a phone number or not. If a phone number is enrolled, the button is hidden, and if not then the user can click on it as normal.