Email authentication bug fixed.
It was discovered during testing that a user with an unauthenticated email address was still able to log in, despite the fact that I had already implemented this feature, so I have no idea why it stopped working.

Nevertheless, the sign in activity has been updated so it again features a check to see if the user has been authenticated. In addition, to help with user flow, the email verification activity now includes a button which takes the user to the main menu if they have authenticated with the resent email.